+++ src/mcmu/__init__.py
@@ -1,6 +1,6 @@
 """Package wide code"""
 
-from logging import getLogger, basicConfig, INFO
+from logging import basicConfig, getLogger
 
 logger = getLogger(__name__)
-basicConfig(format="%(levelname)s:%(name)s %(message)s", level=INFO)
+basicConfig(format="%(levelname)s:%(name)s %(message)s", level=20)